requirements.txt
# Core FastAPI and web framework
fastapi==0.116.2
uvicorn==0.24.0
pydantic==2.11.9
pydantic-settings==2.10.1
python-multipart==0.0.20
aiofiles==23.2.1

# AI/ML Core Dependencies
torch==2.1.0
torchaudio==2.1.0
torchvision==0.16.0
transformers==4.35.0
numpy==1.23.4
scipy==1.10.1
scikit-learn==1.1.3
scikit-image==0.19.3

# TTS and Audio Processing
TTS==0.22.0
librosa==0.9.2
soundfile==0.12.1
sounddevice==0.5.2
pydub==0.25.1
resampy==0.3.1
soxr==1.0.0

# Computer Vision and Image Processing
opencv-python==4.7.0.72
opencv-contrib-python==4.11.0.86
pillow==11.3.0
imageio==2.37.0
imageio-ffmpeg==0.4.7
mediapipe==0.10.21
kornia==0.6.8
kornia_rs==0.1.9

# SadTalker Dependencies
sadtalker-z==0.1.6
face-alignment==1.3.5
facexlib==0.3.0
gfpgan==1.3.8
basicsr==1.4.2
einops==0.8.1

# Text Processing and NLP
nltk==3.9.1
spacy==3.8.7
sentencepiece==0.2.1
tokenizers==0.14.1
huggingface-hub==0.35.0

# Utilities
PyYAML==6.0.2
python-dotenv==1.0.0
requests==2.32.5
tqdm==4.67.1
loguru==0.7.2
coloredlogs==15.0.1
aiohttp==3.12.15

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1
ruff==0.13.1

# Optional: Foundry Local (if using)
# foundry-local-client